fix(onboarding): resolve RSC
